Sketch out visual style of the page.
First, the cat has the name now, Splotch. Has something to do with a
coffee accident. The color scheme is sepia-inspired, to give it a look
and feel of an old drawing.

Splotch's picture occupies most of the first screen space, he's the main
attraction here. Splotch's status is also visible at the bottom, but
it's meant to be very minimal. Footer with the copyright notice and
whatever else I decide to add is pushed out of the viewport by default,
it's not interesting.