Implement a time-dependent smooth noise generator.
This will be later used to introduce some unpredictability into the cat
behavior, for example unprompted mood changes. It is important that the
noise is deterministic and is repeatable, given the same seed and
timestamp, which means that different users loading the page at the same
time should see the same behavior.

The noise smoothing part ensures that the behavior doesn't swing too
wildly from moment to moment. The smoothing algorithm is inspired by 1d
Perlin Noise, except we won't be generating whole bitmaps with it, just
sampling at different points in time.

The underlying randomness source is hash-based, since it's the easiest
way to generate coordinate-dependent repeatable pseudo-random numbers. I
used md5 because it's easily available and reasonably cheap to compute.
I don't need cryptographic safety for what I'm going to use it for.